requirements.txt
streamlit
spacy
numpy
scikit-learn
pdfminer.six
python-docx
fpdf
